fix(session): hook skill telemetry into the actual tool event handler
OpenCode emits tool invocations via the message.part.updated SSE event,
which maps to onToolExecuting → handleToolExecuting — NOT to the
tool_call branch of handleMessagePartCreated (that branch exists for
a different wire format and never fires in practice). Move the
incrementSkillUsage hook to the "new tool call" else branch of
handleToolExecuting so it fires exactly once per real invocation.

Fields in this path: event.toolName is the tool (e.g. "skill"),
event.arguments is part.state.input (e.g. { name: "sentry-fix" }) —
so the fallback chain becomes args.name → args.skill → args.skill_name
(reversed from before; args.name is the canonical field here).
